rag-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.2.5
langchain-community==0.2.5
langchain-openai==0.1.8
langchain-anthropic==0.1.15
langchain-text-splitters==0.2.4
qdrant-client==1.9.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.1
httpx==0.27.0
tiktoken==0.7.0
pypdf==4.2.0
python-multipart==0.0.9
openai==1.109.1
anthropic==0.84.0
pymupdf==1.24.5
python-docx==1.1.2
openpyxl==3.1.4
